requirements.txt
cloudpickle==3.1.1
gym==0.26.2
imageio==2.35.1
isaacgym.egg==info
Jinja2==3.1.6
matplotlib==3.7.5
mujoco_py==2.1.2.14
numpy==1.23.0
open3d==0.19.0
Pillow==11.3.0
pyassimp==5.2.5
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.16.2
setproctitle==1.3.6
tensorboard==2.20.0
tensorboardX==2.6.2.2
tensorboardX==2.6.4
torch==2.4.1
tqdm==4.67.1
urdfpy==0.0.22
wandb==0.21.4
